0.90 to 1.0.2 conversion.

+++ Source/AsteroidHangars/AnimatedConverterBase.cs
@@ -200,7 +200,7 @@ namespace AtHangar
 		protected void produce_heat() 
 		{ 
 			part.temperature += HeatProduction * Rate * vessel.VesselValues.HeatProduction.value * Time.deltaTime;
-			Temperature = part.temperature;
+			Temperature = (float)part.temperature;
 		}
 
 		public void FixedUpdate()
